setting right CategoryController

+++ src/main/java/com/kemsu/sibiryakov/api/Controllers/CategoryController.java
@@ -3,6 +3,7 @@ package com.kemsu.sibiryakov.api.Controllers;
 import com.kemsu.sibiryakov.api.DTOs.CategoryDTO.CategoryAddOneDTO;
 import com.kemsu.sibiryakov.api.DTOs.CategoryDTO.CategoryAddManyDTO;
 import com.kemsu.sibiryakov.api.Entities.Category;
+import com.kemsu.sibiryakov.api.Entities.Emuns.ERole;
 import com.kemsu.sibiryakov.api.Services.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,6 +12,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static com.kemsu.sibiryakov.api.Services.RightsService.checkRight;
 
 @RestController
 @RequestMapping("/api/category")
@@ -22,32 +25,58 @@ public class CategoryController {
     }
 
     @GetMapping("/all")
-    @ResponseStatus(HttpStatus.OK)
-    public List<Category> getAll() {
-        return categoryService.getAllCategory();
+    public ResponseEntity<List<Category>> getAll(@CookieValue(value = "jwt", required = false) String jwt) {
+        if (checkRight(jwt, ERole.ADMINISTRATOR)) {
+            List<Category> categories = categoryService.getAllCategory();
+
+            return categories != null
+                    ? new ResponseEntity<>(categories, HttpStatusCode.valueOf(201))
+                    : new ResponseEntity<>(HttpStatusCode.valueOf(404));
+        } else {
+            return new ResponseEntity<>(HttpStatusCode.valueOf(403));
+        }
     }
 
     @PostMapping("/add")
-    public ResponseEntity<Category> addOneCategory(@RequestBody CategoryAddOneDTO categoryAddOneDTO) {
-        Category category = categoryService.createOneCategory(categoryAddOneDTO);
+    public ResponseEntity<Category> addOneCategory(@RequestBody CategoryAddOneDTO categoryAddOneDTO,
+                                                   @CookieValue(value = "jwt", required = false) String jwt) {
+        if (checkRight(jwt, ERole.ADMINISTRATOR)) {
 
-        return category != null
-                ? new ResponseEntity<>(category, HttpStatusCode.valueOf(201))
-                : new ResponseEntity<>(HttpStatusCode.valueOf(400));
+            Category category = categoryService.createOneCategory(categoryAddOneDTO);
+
+            return category != null
+                    ? new ResponseEntity<>(category, HttpStatusCode.valueOf(201))
+                    : new ResponseEntity<>(HttpStatusCode.valueOf(400));
+        } else {
+            return new ResponseEntity<>(HttpStatusCode.valueOf(403));
+        }
     }
 
     @PostMapping("/addMany")
-    public ResponseEntity<List<Category>> addManyCategory(@RequestBody CategoryAddManyDTO categoryAddManyDTO) {
-        List<Category> categories = categoryService.createManyCategory(categoryAddManyDTO);
+    public ResponseEntity<List<Category>> addManyCategory(@RequestBody CategoryAddManyDTO categoryAddManyDTO,
+                                                          @CookieValue(value = "jwt", required = false) String jwt) {
+        if (checkRight(jwt, ERole.ADMINISTRATOR)) {
 
-        return categories != null
-                ? new ResponseEntity<>(categories, HttpStatusCode.valueOf(201))
-                : new ResponseEntity<>(HttpStatusCode.valueOf(400));
+            List<Category> categories = categoryService.createManyCategory(categoryAddManyDTO);
+
+            return categories != null
+                    ? new ResponseEntity<>(categories, HttpStatusCode.valueOf(201))
+                    : new ResponseEntity<>(HttpStatusCode.valueOf(400));
+        } else {
+            return new ResponseEntity<>(HttpStatusCode.valueOf(403));
+        }
     }
 
     @DeleteMapping("/delete/{id}")
     @ResponseStatus(value = HttpStatus.NO_CONTENT, reason = "Ok")
-    public void deleteCategoryById(@PathVariable Long id) {
-        categoryService.deleteOneCategory(id);
+    public ResponseEntity<?> deleteCategoryById(@PathVariable Long id,
+                                   @CookieValue(value = "jwt", required = false) String jwt) {
+        if (checkRight(jwt, ERole.ADMINISTRATOR)) {
+            categoryService.deleteOneCategory(id);
+
+            return new ResponseEntity<>(HttpStatusCode.valueOf(200));
+        } else {
+            return new ResponseEntity<>(HttpStatusCode.valueOf(403));
+        }
     }
 }
